Initial commit
Update README.md

The Flame Forecast project was created to address the issue of wildfire prediction. Flame Forecast is an intelligent system that can predict how a fire will spread within the first 24 hours after it starts. The goal is to provide timely alerts to people or communities near the fire area. This project focuses on communities that lack access to a network or the Internet. The goal is to develop a device that can send an alert about the extent of the fire without depending on a network. Ensuring the well-being of all people and preventing major disasters is considered of great importance, which is why we are committed to meeting this challenge.

Create Arduino

Create Prototipo

ino App, server-1

Add files via upload

Add fire simulator subrepo

Creado mediante Colaboratory

README structure

MOC added

Introduction redirect